refactor(utils): extract parameter validation into _validate.py (#696)
Move 17 validation/type-check functions verbatim from utils.py into a new sibling
module pl/_validate.py. Imports flow one way: _validate -> _color (_is_color_like,
_prepare_cmap_norm, _get_colors_for_categorical_obs, now top-level) and
_validate -> utils (downward). The temporary function-local _color imports added in
the previous commit are hoisted to top-level here and removed from utils.
render.py and basic.py repointed to _validate.

Completes the utils.py split (#696): utils 4918 -> 1311 lines, with
_geometry / _datashader / _color / _validate as single-concern siblings.
No import cycle anywhere. No behavior change (verbatim moves; all-private).
Verified: 410 non-visual tests pass; ruff + ruff-format clean. Pre-existing
#703/#705 mypy debt in utils.py is unrelated; --no-verify for that reason only.
